+++ extensions/sequence_text/examples/sequence_text_usage.cc
@@ -15,12 +15,20 @@ int main(int argument_count, char** argument_values) {
   try {
     const seqpro::IndexedFasta indexed_fasta = seqpro::IndexedFasta::Open(argument_values[1]);
     seqpro::SequenceTextLayout sequence_text_layout(indexed_fasta);
-    const seqpro::MaterializedSequenceText materialized_text = sequence_text_layout.Materialize();
+    const seqpro::MaterializedSequenceText first_generation_text =
+        sequence_text_layout.Materialize();
 
-    std::cout << "generation: " << materialized_text.layout_generation << '\n'
+    sequence_text_layout.ExcludeTextIntervals(first_generation_text.layout_generation,
+                                               {{1, 2}});
+    sequence_text_layout.Finalize();
+    const seqpro::MaterializedSequenceText second_generation_text =
+        sequence_text_layout.Materialize();
+
+    std::cout << "first_generation\t" << first_generation_text.layout_generation << '\n'
+              << "second_generation\t" << second_generation_text.layout_generation << '\n'
               << "selected sequences: " << sequence_text_layout.sequence_order().size() << '\n'
               << "active runs: " << sequence_text_layout.active_run_count() << '\n'
-              << "text bytes: " << materialized_text.sequence_text_bytes.size() << '\n';
+              << "text bytes: " << second_generation_text.sequence_text_bytes.size() << '\n';
 
     const seqpro::SequenceTextLocation last_location =
         sequence_text_layout.LocateTextPosition(sequence_text_layout.text_size() - 1U);